Interim progress -- works on MacOS

+++ jena/JosekiExtensions/resources/conf/unix/parliament.sh
@@ -1,33 +1,34 @@
-#!/bin/sh
+#!/bin/bash
 
 ######### Find the Parliament installation directory: #########
 cd "`dirname ""$0""`"
 PMNT_DIR="`pwd`"
 cd - > /dev/null
 
-######### User-settable configuration parameters: #########
-export PARLIAMENT_KB_CONFIG_PATH=$PMNT_DIR/ParliamentKbConfig.txt
-export PARLIAMENT_LOG_CONFIG_PATH=$PMNT_DIR/ParliamentLogConfig.txt
 
-MIN_MEM=128m
-MAX_MEM=512m
+######### User-settable configuration parameters: #########
 # Set JETTY_HOST to 0.0.0.0 to make it accessible from other machines on the network:
 JETTY_HOST=localhost
 #JETTY_HOST=0.0.0.0
 JETTY_PORT=8089
+JAVA_HEAP_SIZE=512m
+DAEMON_USER=iemmons
 
-DAEMON_USER=jsmith
+export PARLIAMENT_KB_CONFIG_PATH=$PMNT_DIR/ParliamentKbConfig.txt
+export PARLIAMENT_LOG_CONFIG_PATH=$PMNT_DIR/ParliamentLogConfig.txt
+
 LOG_FILE=$PMNT_DIR/log/jsvc.log
 
 # Uncomment this line to enable remote debugging:
 #DEBUG_ARG=-agentlib:jdwp=transport=dt_socket,address=8000,server=y,suspend=n
+
 
 ######### Error checking & environment detection: #########
 if [ -d "/etc/systemd/system" -a -f "`which systemctl`" ]; then
 	IS_SYSTEM_D="true"
 fi
 
-if [ "$1" != "interactive" -a ! -d "$JAVA_HOME" ]; then
+if [ ! -d "$JAVA_HOME" ]; then
 	echo The JAVA_HOME environment variable is not set.
 	exit 1
 fi
@@ -47,6 +48,7 @@ if [ ! -f "$PARLIAMENT_LOG_CONFIG_PATH" ]; then
 	exit 1
 fi
 
+
 ######### Compute the PID file location: #########
 KB_DIR=`sed -n 's/kbDirectoryPath[ \t]*=\(.*\)$/\1/p' $PARLIAMENT_KB_CONFIG_PATH | tail -n 1 | tr -d '[:space:]'`
 SAVED_DIR=`pwd`
@@ -60,6 +62,7 @@ KB_DIR="`pwd`"
 cd "$SAVED_DIR"
 PID_FILE=$KB_DIR/pid.txt
 
+
 ######### Set up the shared lib path: #########
 # The Java property "java.library.path" below is supposed to take care of these,
 # but sometimes it doesn't work, so set up the shared lib path as well:
@@ -72,31 +75,27 @@ case "`uname -s`" in
 		;;
 esac
 
-######### Set up the command line: #########
-if [ "$1" = 'interactive' ]; then
-	MAIN_CLASS=com.bbn.parliament.jena.jetty.CmdLineJettyServer
-	LOG4J_CONFIG=interactive
-	EXEC='java -server'
-else
-	MAIN_CLASS=com.bbn.parliament.jena.jetty.JettyDaemon
-	LOG4J_CONFIG=daemon
-	EXEC="$PMNT_DIR/bin/jsvc -jvm server -showversion -home ""$JAVA_HOME"" -cwd $PMNT_DIR"
-	if [ -n "$DAEMON_USER" ]; then
-		EXEC="$EXEC -user ""$DAEMON_USER"""
-	fi
-	EXEC="$EXEC -outfile $LOG_FILE -errfile &1 -pidfile $PID_FILE -procname Parliament"
-	# Uncomment this line to enable verbose jsvc output:
-	#EXEC="$EXEC -debug"
-fi
 
-EXEC="$EXEC -Xms$MIN_MEM -Xmx$MAX_MEM -cp $CLASSPATH:$PMNT_DIR/lib/*"
+######### Set up the command line: #########
+MAIN_CLASS=com.bbn.parliament.jena.jetty.JettyDaemon
+EXEC="$PMNT_DIR/bin/jsvc -jvm server -showversion -home ""$JAVA_HOME"" -cwd $PMNT_DIR"
+if [ -n "$DAEMON_USER" ]; then
+	EXEC="$EXEC -user ""$DAEMON_USER"""
+fi
+EXEC="$EXEC -outfile $LOG_FILE -errfile &1 -pidfile $PID_FILE -procname Parliament"
+
+# Uncomment this line to enable verbose jsvc output:
+#EXEC="$EXEC -debug"
+
+EXEC="$EXEC -Xmx$JAVA_HEAP_SIZE -cp $CLASSPATH:$PMNT_DIR/lib/*"
 EXEC="$EXEC -Djava.library.path=$PMNT_DIR/bin -Dcom.sun.management.jmxremote"
-EXEC="$EXEC -Dlog4j.configuration=file:$PMNT_DIR/conf/log4j.$LOG4J_CONFIG.properties"
+EXEC="$EXEC -Dlog4j.configuration=file:$PMNT_DIR/conf/log4j.daemon.properties"
 EXEC="$EXEC -Djetty.host=$JETTY_HOST -Djetty.port=$JETTY_PORT"
 EXEC="$EXEC -DjettyConfig=$PMNT_DIR/conf/jetty.xml"
 if [ -n "$DEBUG_ARG" ]; then
 	EXEC="$EXEC ""$DEBUG_ARG"""
 fi
+
 
 ######### Debugging statements: #########
 # echo EXEC = $EXEC [-stop] $MAIN_CLASS
@@ -104,54 +103,49 @@ fi
 # echo PID_FILE = $PID_FILE
 
 
-
-
-
-
-
-
-
-
-
-
-
 ######### Usage of this script: #########
 function printUsage {
+	echo ""
+	echo "Usage: $(basename $0) {start|stop|restart|install|uninstall}"
+	echo ""
+	echo "where:"
+	echo ""
+	echo "   'start' runs Parliament as a detached process"
+	echo ""
+	echo "   'stop' stops the detached Parliament process"
+	echo ""
+	echo "   'restart' stops and then immediately starts Parliament"
+	echo ""
 	if [ "$IS_SYSTEM_D" = "true" ]; then
-		echo "Usage: $(basename $0) {install|uninstall|interactive}"
+		echo "   'install' sets Parliament up as a systemd service, so"
+		echo "      it can be controlled via the systemctl command"
 		echo ""
-		echo "where:"
-		echo ""
-		echo "   install will create Parliament as a system-d service, after which Parliament can be controlled via the systemctl command"
-		echo ""
-		echo "   uninstall will remove the Parliament system-d service"
-		echo ""
-		echo "   interactive will start Parliament running as an interactive process in the current shell"
+		echo "   'uninstall' removes the Parliament systemd service"
+		echo "      definition"
 	else
-		echo "Usage: $(basename $0) {start|stop|restart|interactive}"
+		echo "   'install' [not available on this platform]"
 		echo ""
-		echo "where:"
-		echo ""
-		echo "   start will start Parliament running as a detached process"
-		echo ""
-		echo "   stop will stop a detached Parliament process"
-		echo ""
-		echo "   restart will stop and then immediately re-start a detached Parliament process"
-		echo ""
-		echo "   interactive will start Parliament running as an interactive process in the current shell"
+		echo "   'uninstall' [not available on this platform]"
 	fi
+	echo ""
 	exit 3
 }
 
+
 ######### Create systemd service file: #########
 # Takes path of file as an argument
-function writeSystemDServiceFile {
+function installSystemDService {
+	if [ -z "$DAEMON_USER" ]; then
+		echo 'Please set the DAEMON_USER variable (near the top of the script).'
+		exit 1
+	fi
+
 	if [ -f "$1" ]; then
 		rm $1
 	fi
 
 	echo '[Unit]' >> $1
-	echo 'Description=Parliament Service' >> $1
+	echo 'Description=Parliament Semantic Graph Service' >> $1
 	echo 'After=network.target' >> $1
 	echo 'StartLimitIntervalSec=0' >> $1
 	echo '' >> $1
@@ -175,55 +169,40 @@ function writeSystemDServiceFile {
 	echo 'WantedBy=multi-user.target' >> $1
 }
 
-######### Create systemd service file: #########
-function installSystemDService {
-	if [ -z "$DAEMON_USER" ]; then
-		echo 'Please set the DAEMON_USER variable (near the top of the script).'
-		exit 1
-	fi
-
-	writeSystemDServiceFile './parliament.service.txt'
-	#writeSystemDServiceFile '/etc/systemd/system/parliament.service'
-}
-
-
-
-
-
-
-
-
-
-
-
-
 
 ######### Execute: #########
-if [ "$IS_SYSTEM_D" = "true" ]; then
-	echo "Running on a System-D operating system"
-else
-	echo "This operating system is not System-D"
-fi
-if [ "$1" = "interactive" ]; then
+if [ "$1" = "start" ]; then
 	$EXEC $MAIN_CLASS
-elif [ "$1" = "start" -a "$IS_SYSTEM_D" != "true" ]; then
-	$EXEC $MAIN_CLASS
-elif [ "$1" = "stop" -a "$IS_SYSTEM_D" != "true" ]; then
+elif [ "$1" = "stop" ]; then
 	if [ -f "$PID_FILE" ]; then
 		$EXEC -stop $MAIN_CLASS
 	else
 		echo 'The Parliament daemon is not running.'
 		exit 1
 	fi
-elif [ "$1" = "restart" -a "$IS_SYSTEM_D" != "true" ]; then
+elif [ "$1" = "restart" ]; then
 	if [ -f "$PID_FILE" ]; then
 		$EXEC -stop $MAIN_CLASS
 	fi
 	$EXEC $MAIN_CLASS
 elif [ "$1" = "install" -a "$IS_SYSTEM_D" = "true" ]; then
+	installSystemDService '/etc/systemd/system/parliament.service'
+	# Does this next command need to be preceded by'systemctl start parliament'?
+	systemctl enable parliament
+	echo 'You can now control Parliament via the command'
+	echo '   systemctl \{ start | stop | restart \} parliament'
 elif [ "$1" = "uninstall" -a "$IS_SYSTEM_D" = "true" ]; then
+	systemctl stop parliament
+	systemctl disable parliament
+	rm '/etc/systemd/system/parliament.service'
+	echo 'Parliament has been uninstalled as a systemd service'
 elif [ "$1" = "test" ]; then
-	installSystemDService
+	if [ "$IS_SYSTEM_D" = "true" ]; then
+		echo "Running on a systemd operating system"
+	else
+		echo "This is not a systemd operating system"
+	fi
+	installSystemDService './parliament.service.txt'
 else
 	printUsage
 fi
